requirements.txt
click==8.1.6
gensim==4.3.1
joblib==1.3.1
mysql-connector-python==8.1.0
nltk==3.8.1
numpy==1.25.2
pika==1.3.2
protobuf==4.21.12
python-dotenv==1.0.0
regex==2023.6.3
scipy==1.11.1
smart-open==6.3.0
tqdm==4.65.0